add : MemberService 추상화
