+++ src/test/java/com/jejeong/apipractice/sevice/member/MemberServiceTest.java
@@ -21,50 +21,51 @@ import static org.mockito.BDDMockito.given;
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
 
-    private MemberService memberService;
+  private MemberService memberService;
 
-    @Mock
-    private MemberRepository memberRepository;
+  @Mock
+  private MemberRepository memberRepository;
 
-    @BeforeEach
-    void beforeEach() {
-        memberService = new MemberService(memberRepository);
-    }
+  @BeforeEach
+  void beforeEach() {
+    memberService = new GeneralMemberService(memberRepository);
+  }
 
-    @Test
-    void loadUserByUserEmail() {
-        // given
-        Member member = createMember();
-        given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
+  @Test
+  void loadUserByUserEmail() {
+    // given
+    Member member = createMember();
+    given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
 
-        // when
-        MemberDto memberDto = memberService.loadUserByUserEmail(member.getEmail());
+    // when
+    MemberDto memberDto = memberService.loadUserByUserEmail(member.getEmail());
 
-        // then
-        assertThat(memberDto.getEmail()).isEqualTo(member.getEmail());
+    // then
+    assertThat(memberDto.getEmail()).isEqualTo(member.getEmail());
 
-    }
+  }
 
-    @Test
-    @DisplayName("회원이 없는 경우 에러를 내뱉는다.")
-    void test3() {
-        // given
-        given(memberRepository.findByEmail(anyString())).willReturn(Optional.empty());
+  @Test
+  @DisplayName("회원이 없는 경우 에러를 내뱉는다.")
+  void test3() {
+    // given
+    given(memberRepository.findByEmail(anyString())).willReturn(Optional.empty());
 
-        // then
-        assertThatThrownBy(() -> memberService.loadUserByUserEmail(anyString())).isInstanceOf(IllegalArgumentException.class);
-    }
+    // then
+    assertThatThrownBy(() -> memberService.loadUserByUserEmail(anyString())).isInstanceOf(
+        IllegalArgumentException.class);
+  }
 
-    @Test
-    void deleteUserByUserEmail() {
-        // given
-        Member member = createMember();
-        given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
+  @Test
+  void deleteUserByUserEmail() {
+    // given
+    Member member = createMember();
+    given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
 
-        // when
-        memberService.deleteUserByUserEmail(member.getEmail());
+    // when
+    memberService.deleteUserByUserEmail(member.getEmail());
 
-        // then
-        assertThat(member.getEmail()).isEqualTo("******");
-    }
+    // then
+    assertThat(member.getEmail()).isEqualTo("******");
+  }
 }
